refactor: Move functions in check_map_contents() to new file identifier.c and split the function (#25)
